steamrollArray Done
Not very happy with this one. I couldn't think of a good way to do this,
I ended up having to use one of the methods I saw online using
concat.apply and then looping through the leftovers for a final check. I
spent way too much time on this and am not happy with the end result :(